# apps/ade-api/src/ade_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Automatic Data Extractor API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/rbac/permissions": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List permissions",
        "operationId": "list_permissions_api_v1_rbac_permissions_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ScopeType",
              "description": "Scope to filter permissions by",
              "default": "global"
            },
            "description": "Scope to filter permissions by"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Total"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/rbac/roles": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List role definitions",
        "operationId": "list_roles_api_v1_rbac_roles_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ScopeType",
              "description": "Scope to filter roles by",
              "default": "global"
            },
            "description": "Scope to filter roles by"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Total"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "rbac"
        ],
        "summary": "Create a role",
        "operationId": "create_role_api_v1_rbac_roles_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/rbac/roles/{role_id}": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "Retrieve a role definition",
        "operationId": "read_role_api_v1_rbac_roles__role_id__get",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Role identifier",
              "title": "Role Id"
            },
            "description": "Role identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "rbac"
        ],
        "summary": "Update an existing role",
        "operationId": "update_role_api_v1_rbac_roles__role_id__patch",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Role identifier",
              "title": "Role Id"
            },
            "description": "Role identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "delete": {
        "tags": [
          "rbac"
        ],
        "summary": "Delete a role",
        "operationId": "delete_role_api_v1_rbac_roles__role_id__delete",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Role identifier",
              "title": "Role Id"
            },
            "description": "Role identifier"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/rbac/role-assignments": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List role assignments (admin view)",
        "operationId": "list_assignments_api_v1_rbac_role_assignments_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ScopeType",
              "description": "Scope to filter assignments by",
              "default": "global"
            },
            "description": "Scope to filter assignments by"
          },
          {
            "name": "scope_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scope ID (required when scope=workspace)",
              "title": "Scope Id"
            },
            "description": "Scope ID (required when scope=workspace)"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by user id",
              "title": "User Id"
            },
            "description": "Filter by user id"
          },
          {
            "name": "role_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by role id",
              "title": "Role Id"
            },
            "description": "Filter by role id"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Total"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/users/{user_id}/roles": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List global roles assigned to a user",
        "operationId": "list_user_roles_api_v1_users__user_id__roles_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User identifier",
              "title": "User Id"
            },
            "description": "User identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermissionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PermissionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "has_next",
          "has_previous"
        ],
        "title": "PermissionPage",
        "description": "Paginated permission registry response."
      },
...
      "ScopeType": {
        "type": "string",
        "enum": [
          "global",
          "workspace"
        ],
        "title": "ScopeType",
        "description": "Scopes supported by RBAC-aware resources."
      },
...
      "UserRolesEnvelope": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/UserRoleSummary"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "roles"
        ],
        "title": "UserRolesEnvelope",
        "description": "Envelope for listing all global roles for a user."
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ade_session",
        "description": "Browser session cookie issued after interactive sign-in."
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer access token returned by ADE or an identity provider."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key for service integrations."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "security": [
    {
      "SessionCookie": []
    },
    {
      "HTTPBearer": []
    },
    {
      "APIKeyHeader": []
    }
  ]
}